move translation to compontends to AbeillePanel and remove FormPanelI18N.java
